backend/src/python/requirements.txt
pymongo
python-dotenv
sentence-transformers
faiss-cpu
PyPDF2
transformers
torch
OpenAI